dataflow/pyproject.toml
[tool.poetry]
name = "dataflow"
version = "0.1.0"
description = ""
authors = ["Raynor <jonggu.woo@fitpet.co.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
black = "^25.1.0"
pandas = "^2.2.3"
jupyter = "^1.1.1"
xmltodict = "^0.14.2"
python-dotenv = "^1.0.1"
thread-manager-py = "^0.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py312']
skip-string-normalization = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
